% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_builder.R
\name{background_builder}
\alias{background_builder}
\title{Background Builder for Species Distribution Modelling}
\usage{
background_builder(distrib = distrib, terres = terres,
  wrld_simpl = wrld_simpl, ref_rast = ref_rast, no_biome = FALSE)
}
\arguments{
\item{distrib}{Distribution data with "Longitude" and "Latitude" colnames}

\item{terres}{The terrestrial biomes layer from Olson et al.}

\item{wrld_simpl}{The simple country/ continent boundaries from maptools}

\item{ref_rast}{Reference raster for grid cell processing}

\item{no_biome}{Flag to turn off the biome part and just return continent level backgrounds}
}
\value{
A \code{raster} file that can be used for sampling background
}
\description{
Background Builder for Species Distribution Modelling
}
