% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_builder.R
\name{bg_weighting}
\alias{bg_weighting}
\title{Background weighting function}
\usage{
bg_weighting(bg_ras = bg_ras, weight = 1e-06, spp.data = spp.data)
}
\arguments{
\item{bg_ras}{The background raster to sample across}

\item{weight}{The value for background cells}

\item{spp.data}{The presence data - must include a column of 0/1 for presence background samples}
}
\value{
who knows
}
\description{
Background weighting function
}
