% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/che.R
\name{che_model}
\alias{che_model}
\title{Perform convex hulls models}
\usage{
che_model(
  spp = "spp",
  parallel = TRUE,
  model.data = model.data,
  n.cores = detectCores() - 1
)
}
\arguments{
\item{spp}{Name of species}

\item{parallel}{T/F flag for using parallel - must be TRUE for now}

\item{model.data}{A model.data list of formatted data from \code{prepare_data} function.}
}
\value{
Modelled outputs
}
\description{
Wrapper function for the CHE modelling method
}
