% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insect_dist}
\alias{insect_dist}
\title{Distribution data for 22 globally invasive insects species}
\format{A data frame with 11554 rows and 4 variables:
\describe{
  \item{Species}{Species name}
  \item{Longitude}{Longitude in WGS84}
  \item{Latitiude}{Latitude in WGS84}
  \item{Range}{Points are from native or invasive range}
  ...
}}
\source{
\url{https://doi.org/10.6084/m9.figshare.4537922.v2}
}
\usage{
insect_dist
}
\description{
A dataset containing the native and invasive distributions for 22 insect species.
}
\references{
Hill MP, Gallardo B, Terblanche JS (2017) A global assessment of climatic niche shifts, and human influence, in insect invasions Global Ecology & Biogeography 26, 679-689
}
\keyword{datasets}
