% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/che.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for use in the CHE model}
\usage{
prepare_data(
  rasters = rasters,
  mask_raster = mask,
  predict.to = predict.to,
  spp.dist = spp.dist,
  w.val = 1e-06,
  bg.sample = 50000
)
}
\arguments{
\item{rasters}{Stack of rasters that are bioclim variables and cover extent of study (global fine)}

\item{mask_raster}{Background mask, usualyl created using \code{background_builder}}

\item{predict.to}{Raster stack where the model will be predicted/ projected to}

\item{spp.dist}{X & Y species points}

\item{w.val}{Weighting value for background}

\item{bg.sample}{Size of background samples (default 50000)}
}
\value{
A list of data objects to be used with che model.
}
\description{
Prepare data for use in the CHE model
}
