% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_bagging.R
\name{rb}
\alias{rb}
\title{Range Bagging algorithm for Species Distribution Modelling}
\usage{
rb(x, v = 100, d = 1, p = 0.5, bioclim = TRUE)
}
\arguments{
\item{x}{covariate data at observations}

\item{v}{number of votes}

\item{d}{dimension of ranges to bag}

\item{bioclim}{Flag to indicate whether to select 1 precipitation and 1 temperature variable at at time}
}
\value{
A \code{list} of length \code{v} containing the individual models
}
\description{
Basic version of the range bagging algorithm
}
