% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_bagging.R
\name{rb.test}
\alias{rb.test}
\title{Range Bagging prediction}
\usage{
rb.test(models, x.new)
}
\arguments{
\item{models}{a list of models returned by function rb}

\item{x.new}{covariate data at new points}
}
\value{
A \code{list} of length \code{v} containing the individual models
}
\description{
Test function for a point to determine the fraction of bags in which the point falls
}
\references{
Drake, J. M. (2015). Range bagging: a new method for ecological niche modelling from presence-only data. Journal of The Royal Society Interface, 12(107):20150086
}
\author{
John Drake
}
