% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/che.R
\name{runCHE}
\alias{runCHE}
\title{Convex Hull Ensembles}
\usage{
runCHE(j)
}
\arguments{
\item{j}{Counter for bioclim raster position}
}
\value{
A \code{list} comprising two stacks of rasters
}
\description{
Convex Hull Ensembles
}
