% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/che.R
\name{sum_stack}
\alias{sum_stack}
\title{Stacking function for che outputs}
\usage{
sum_stack(stackoflayers, keep_layers = keep_layers)
}
\arguments{
\item{stackoflayers}{Raster layers of che results}

\item{keep_layers}{Vector of layer names to keep}
}
\value{
A single layer
}
\description{
Stacking function for che outputs
}
